+++ code/include/util.h
@@ -2,11 +2,12 @@
 #define UTIL_H
 
 #include <stdio.h>
+#include <math.h>
 #include <bits/stdc++.h> 
 using namespace std;
 
 // Bark-frequency conversion
-float bark_to_hertz(float bark);
-float hertz_to_bark(float hertz); 
+double bark_to_hertz(double bark);
+double hertz_to_bark(double hertz); 
 
 #endif
